Update README: access_db feature is required for access entries
The access entries only work when the access_db feature is also enabled.
This matches the traditional behaviour when Puppet is not used since the
mere existance of a file /etc/mail/access does not cause Sendmail to use
it.

The README has been updated to note the pitfall of forgetting to add the
feature. This closes #12.